+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.29dev
 
+* [Feature] Supported `save_profiling_data` option with file path
+
 ## 0.0.28 (2023-04-28)
 
 *No changes. Testing the release workflow*